+++ lib/parallel_tests.rb
@@ -44,6 +44,8 @@ module ParallelTests
 
     def stop_all_processes
       pids.all.each { |pid| Process.kill(:INT, pid) }
+    rescue Errno::ESRCH, Errno::EPERM
+      # Process already terminated, do nothing
     end
 
     # copied from http://github.com/carlhuda/bundler Bundler::SharedHelpers#find_gemfile