+++ lib/parallel_tests/tasks.rb
@@ -48,6 +48,9 @@ module ParallelTests
         activate_pipefail = "set -o pipefail"
         remove_ignored_lines = %{(grep -v #{Shellwords.escape(ignore_regex)} || true)}
 
+        # remove nil values (ex: #purge_before_load returns nil)
+        command.compact!
+
         if system('/bin/bash', '-c', "#{activate_pipefail} 2>/dev/null")
           shell_command = "#{activate_pipefail} && (#{Shellwords.shelljoin(command)}) | #{remove_ignored_lines}"
           ['/bin/bash', '-c', shell_command]
@@ -125,6 +128,24 @@ module ParallelTests
         command
       end
 
+      def configured_databases
+        return [] unless defined?(ActiveRecord) && rails_61_or_greater?
+
+        @@configured_databases ||= ActiveRecord::Tasks::DatabaseTasks.setup_initial_database_yaml
+      end
+
+      def for_each_database(&block)
+        # Use nil to represent all databases
+        block&.call(nil)
+
+        # skip if not rails or old rails version
+        return if !defined?(ActiveRecord::Tasks::DatabaseTasks) || !ActiveRecord::Tasks::DatabaseTasks.respond_to?(:for_each)
+
+        ActiveRecord::Tasks::DatabaseTasks.for_each(configured_databases) do |name|
+          block&.call(name)
+        end
+      end
+
       private
 
       def rails_7_or_greater?
@@ -145,25 +166,33 @@ namespace :parallel do
     ParallelTests::Tasks.run_in_parallel(ParallelTests::Tasks.suppress_schema_load_output(command), args)
   end
 
-  desc "Create test databases via db:create --> parallel:create[num_cpus]"
-  task :create, :count do |_, args|
-    ParallelTests::Tasks.run_in_parallel(
-      [$0, "db:create", "RAILS_ENV=#{ParallelTests::Tasks.rails_env}"],
-      args
-    )
+  ParallelTests::Tasks.for_each_database do |name|
+    task_name = 'create'
+    task_name += ":#{name}" if name
+    desc "Create test#{" #{name}" if name} database via db:#{task_name} --> parallel:#{task_name}[num_cpus]"
+    task task_name.to_sym, :count do |_, args|
+      ParallelTests::Tasks.run_in_parallel(
+        [$0, "db:#{task_name}", "RAILS_ENV=#{ParallelTests::Tasks.rails_env}"],
+        args
+      )
+    end
   end
 
-  desc "Drop test databases via db:drop --> parallel:drop[num_cpus]"
-  task :drop, :count do |_, args|
-    ParallelTests::Tasks.run_in_parallel(
-      [
-        $0,
-        "db:drop",
-        "RAILS_ENV=#{ParallelTests::Tasks.rails_env}",
-        "DISABLE_DATABASE_ENVIRONMENT_CHECK=1"
-      ],
-      args
-    )
+  ParallelTests::Tasks.for_each_database do |name|
+    task_name = 'drop'
+    task_name += ":#{name}" if name
+    desc "Drop test#{" #{name}" if name} database via db:#{task_name} --> parallel:#{task_name}[num_cpus]"
+    task task_name.to_sym, :count do |_, args|
+      ParallelTests::Tasks.run_in_parallel(
+        [
+          $0,
+          "db:#{task_name}",
+          "RAILS_ENV=#{ParallelTests::Tasks.rails_env}",
+          "DISABLE_DATABASE_ENVIRONMENT_CHECK=1"
+        ],
+        args
+      )
+    end
   end
 
   desc "Update test databases by dumping and loading --> parallel:prepare[num_cpus]"
@@ -190,12 +219,16 @@ namespace :parallel do
   end
 
   # when dumping/resetting takes too long
-  desc "Update test databases via db:migrate --> parallel:migrate[num_cpus]"
-  task :migrate, :count do |_, args|
-    ParallelTests::Tasks.run_in_parallel(
-      [$0, "db:migrate", "RAILS_ENV=#{ParallelTests::Tasks.rails_env}"],
-      args
-    )
+  ParallelTests::Tasks.for_each_database do |name|
+    task_name = 'migrate'
+    task_name += ":#{name}" if name
+    desc "Update test#{" #{name}" if name} database via db:#{task_name} --> parallel:#{task_name}[num_cpus]"
+    task task_name.to_sym, :count do |_, args|
+      ParallelTests::Tasks.run_in_parallel(
+        [$0, "db:#{task_name}", "RAILS_ENV=#{ParallelTests::Tasks.rails_env}"],
+        args
+      )
+    end
   end
 
   desc "Rollback test databases via db:rollback --> parallel:rollback[num_cpus]"
@@ -207,16 +240,24 @@ namespace :parallel do
   end
 
   # just load the schema (good for integration server <-> no development db)
-  desc "Load dumped schema for test databases via db:schema:load --> parallel:load_schema[num_cpus]"
-  task :load_schema, :count do |_, args|
-    command = [
-      $0,
-      ParallelTests::Tasks.purge_before_load,
-      "db:schema:load",
-      "RAILS_ENV=#{ParallelTests::Tasks.rails_env}",
-      "DISABLE_DATABASE_ENVIRONMENT_CHECK=1"
-    ]
-    ParallelTests::Tasks.run_in_parallel(ParallelTests::Tasks.suppress_schema_load_output(command), args)
+  ParallelTests::Tasks.for_each_database do |name|
+    rails_task = 'db:schema:load'
+    rails_task += ":#{name}" if name
+
+    task_name = 'load_schema'
+    task_name += ":#{name}" if name
+
+    desc "Load dumped schema for test#{" #{name}" if name} database via #{rails_task} --> parallel:#{task_name}[num_cpus]"
+    task task_name.to_sym, :count do |_, args|
+      command = [
+        $0,
+        ParallelTests::Tasks.purge_before_load,
+        rails_task,
+        "RAILS_ENV=#{ParallelTests::Tasks.rails_env}",
+        "DISABLE_DATABASE_ENVIRONMENT_CHECK=1"
+      ]
+      ParallelTests::Tasks.run_in_parallel(ParallelTests::Tasks.suppress_schema_load_output(command), args)
+    end
   end
 
   # load the structure from the structure.sql file