+++ spec/fixtures/rails60/db/migrate/20210325101904_create_users.rb
@@ -1,9 +1,0 @@
-class CreateUsers < ActiveRecord::Migration[6.0]
-  def change
-    create_table :users do |t|
-      t.string :name
-
-      t.timestamps
-    end
-  end
-end