+++ spec/parallel_tests/cli_spec.rb
@@ -57,8 +57,24 @@ describe ParallelTests::CLI do
 
     it "parses --verbose-command" do
       expect(call(['test', '--verbose-command'])).to eq(
-        defaults.merge(verbose_command: true)
+        defaults.merge(verbose_process_command: true, verbose_rerun_command: true)
       )
+    end
+
+    it "parses --verbose-process-command" do
+      expect(call(['test', '--verbose-process-command'])).to eq(
+        defaults.merge(verbose_process_command: true)
+      )
+    end
+
+    it "parses --verbose-rerun-command" do
+      expect(call(['test', '--verbose-rerun-command'])).to eq(
+        defaults.merge(verbose_rerun_command: true)
+      )
+    end
+
+    it "parses --failure-exit-code" do
+      expect(call(["test", "--failure-exit-code", "42"])).to eq(defaults.merge(failure_exit_code: 42))
     end
 
     it "parses --quiet" do
@@ -76,6 +92,10 @@ describe ParallelTests::CLI do
     it "parses --first-is-1" do
       expect(call(["test", "--first-is-1"]))
         .to eq(defaults.merge(first_is_1: true))
+    end
+
+    it "parses allow-duplicates" do
+      expect(call(["test", "--allow-duplicates"])).to eq(defaults.merge(allow_duplicates: true))
     end
 
     context "parse only-group" do
@@ -231,10 +251,10 @@ describe ParallelTests::CLI do
         it_prints_nothing_about_rerun_commands(verbose: false)
       end
 
-      context "with verbose command" do
+      context "with verbose rerun command" do
         it "prints command if there is a failure" do
           expect do
-            subject.send(:report_failure_rerun_commmand, single_failed_command, verbose_command: true)
+            subject.send(:report_failure_rerun_commmand, single_failed_command, verbose_rerun_command: true)
           end.to output("\n\nTests have failed for a parallel_test group. Use the following command to run the group again:\n\nTEST_ENV_NUMBER= PARALLEL_TEST_GROUPS= foo\n").to_stdout
         end
       end
@@ -355,6 +375,45 @@ describe ParallelTests::CLI do
         expect(subject).to receive(:run_tests).once.with(['ccc', 'ddd'], 0, 1, options).and_return(results)
         expect(subject).to receive(:run_tests).once.with(['eee', 'fff'], 1, 1, options).and_return(results)
         subject.run(['test', '-n', '3', '--only-group', '2,3', '-t', 'my_test_runner'])
+      end
+    end
+
+    context 'when --allow-duplicates' do
+      let(:results) { { stdout: "", exit_status: 0 } }
+      let(:processes) { 2 }
+      let(:common_options) do
+        { files: ['test'], allow_duplicates: true, first_is_1: false }
+      end
+      before do
+        allow(subject).to receive(:puts)
+        expect(subject).to receive(:load_runner).with("my_test_runner").and_return(ParallelTests::MyTestRunner::Runner)
+        allow(ParallelTests::MyTestRunner::Runner).to receive(:test_file_name).and_return("test")
+        expect(subject).to receive(:report_results).and_return(nil)
+      end
+
+      before do
+        expect(ParallelTests::MyTestRunner::Runner).to receive(:tests_in_groups).and_return(
+          [
+            ['foo'],
+            ['foo'],
+            ['bar']
+          ]
+        )
+      end
+
+      it "calls run_tests with --only-group" do
+        options = common_options.merge(count: processes, only_group: [2, 3], group_by: :filesize)
+        expect(subject).to receive(:run_tests).once.with(['foo'], 0, 1, options).and_return(results)
+        expect(subject).to receive(:run_tests).once.with(['bar'], 1, 1, options).and_return(results)
+        subject.run(['test', '-n', processes.to_s, '--allow-duplicates', '--only-group', '2,3', '-t', 'my_test_runner'])
+      end
+
+      it "calls run_tests with --first-is-1" do
+        options = common_options.merge(count: processes, first_is_1: true)
+        expect(subject).to receive(:run_tests).once.with(['foo'], 0, processes, options).and_return(results)
+        expect(subject).to receive(:run_tests).once.with(['foo'], 1, processes, options).and_return(results)
+        expect(subject).to receive(:run_tests).once.with(['bar'], 2, processes, options).and_return(results)
+        subject.run(['test', '-n', processes.to_s, '--first-is-1', '--allow-duplicates', '-t', 'my_test_runner'])
       end
     end
   end